Infosys_springboard/requirements.txt
langgraph
langsmith
langchain-core
google-genai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
google-generativeai
langgraph-checkpoint-sqlite
flask
